refactor and fix codesmell
