+++ common/components/elements/CodeBlock.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 'use client'
 
 import dynamic from 'next/dynamic'
@@ -13,6 +14,10 @@ import tsx from 'react-syntax-highlighter/dist/cjs/languages/prism/tsx'
 import typescript from 'react-syntax-highlighter/dist/cjs/languages/prism/typescript'
 import { a11yDark as themeColor } from 'react-syntax-highlighter/dist/cjs/styles/prism'
 import { useCopyToClipboard } from 'usehooks-ts'
+
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
+/* eslint-disable @typescript-eslint/no-unused-vars */
 
 const languages = {
   javascript: 'javascript',
