+++ common/components/layouts/sidebar/MenuItem.tsx
@@ -8,7 +8,7 @@ import { BsArrowRightShort as ExternalLinkIcon } from 'react-icons/bs'
 
 import { MenuItemProps } from '@/common/types/menu'
 
-import { useMenu } from '@/context/menu'
+import { useMenu } from '@/stores/menu'
 
 export default function MenuItem({ title, href, icon, onClick, className = '', children }: MenuItemProps) {
   //   const { hideNavbar } = useContext(MenuContext);
