+++ modules/dashboard/components/PageSpeed/SpeedSection.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react'
 
 import CircleProgress from '@/common/components/elements/CircleProgress'
