+++ stores/blog-view.tsx
